* main_(UMS): (23 commits)
  ConnectedRenderers cleanup last. (#3594)
  fixed setting correct UMS profile parameter (#3595)
  Updated changelog
  Sync translations with Crowdin
  Fix TV series never being found locally by IMDb ID (#3576)
  Update typescript-eslint monorepo to v5.42.1 (#3591)
  Update dependency com.github.spotbugs:spotbugs-maven-plugin to v4.7.3.0 (#3592)
  react web player renderer review (#3590)
  Update dependency @types/jest to v29.2.2 (#3578)
  Update dependency pako to v2.1.0 (#3589)
  Update dependency @types/node to v18 (#3572)
  Confs (#3577)
  Update Mantine monorepo packages to v5.7.1 (#3587)
  Update dependency com.github.spotbugs:spotbugs-maven-plugin to v4.7.2.2 (#3588)
  Update dependency eslint to v8.27.0 (#3586)
  Update Mantine monorepo packages to v5.7.0 (#3583)
  Update dependency com.github.junrar:junrar to v7.5.4 (#3581)
  Update dependency hls.js to v1.2.5 (#3582)
  Added context to the term Update (#3585)
  Update dependency @types/lodash to v4.14.188 (#3580)
  ...

# Conflicts:
#	src/main/java/net/pms/dlna/DbIdResourceLocator.java

+++ src/main/java/net/pms/image/JPEGSubsamplingNotation.java
@@ -1,19 +1,18 @@
 /*
  * This file is part of Universal Media Server, based on PS3 Media Server.
  *
- * This program is a free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; version 2
- * of the License only.
+ * This program is a free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; version 2 of the License only.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+ * details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ * You should have received a copy of the GNU General Public License along with
+ * this program; if not, write to the Free Software Foundation, Inc., 51
+ * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 package net.pms.image;
 
@@ -114,12 +113,12 @@ public class JPEGSubsamplingNotation implements Serializable {
 
 		if (
 			directory.getInteger(JpegDirectory.TAG_NUMBER_OF_COMPONENTS) == null ||
-			directory.getInteger(JpegDirectory.TAG_NUMBER_OF_COMPONENTS).intValue() == 0
+			directory.getInteger(JpegDirectory.TAG_NUMBER_OF_COMPONENTS) == 0
 		) {
 			return new JPEGSubsamplingNotation(Double.NaN, Double.NaN, Double.NaN);
 		}
 
-		int numComponents = directory.getInteger(JpegDirectory.TAG_NUMBER_OF_COMPONENTS).intValue();
+		int numComponents = directory.getInteger(JpegDirectory.TAG_NUMBER_OF_COMPONENTS);
 		int luminanceIdx = -1;
 		JpegComponent[] components = new JpegComponent[numComponents];
 		for (int i = 0; i < numComponents; i++) {
